Added Lambda function to stop EC2 instances and updated DynamoDB table with its status.
A new Lambda function named LambdaAutoStopper has been introduced. This function is triggered when an EC2 instance enters an alarm state, which stops the instances and updates a DynamoDB table with the instance state. An ARN environment variable for this function was also added to the CloudFormation template to link the two lambdas together.